[BugFix] return BIGINT instead of INT out of match_activities (#42)
- [x] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related
issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

<!-- (Please describe the changes you have made. In which scenarios will
this bug be triggered and what measures have you taken to fix the bug?)
-->

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function

+++ be/src/exprs/celonis/match_activities.cpp
@@ -14,7 +14,7 @@ ColumnPtr _celonis_match_activities_impl(FunctionContext* context, const Column&
                                          const NullColumn::Container* null_offsets, const SliceHashSet& nodes) {
     const size_t num_array = offsets.size() - 1;
     auto offsets_ptr = offsets.get_data().data();
-    auto result_column = Int32Column::create(num_array, 0);
+    auto result_column = Int64Column::create(num_array, 0);
     int* result_data = result_column->get_data().data();
     using ValueType = RunTimeCppType<TYPE_VARCHAR>;
     auto elements_ptr = (const ValueType *) (elements.raw_data());
@@ -57,7 +57,7 @@ StatusOr<ColumnPtr> CelonisMatchActivitiesFunctions::celonis_match_activities(Fu
         std::stringstream error;
         error << "unexpected null activities array as input to celonis_match_activities" << std::endl;
         context->set_error(error.str().c_str());
-        return Int32Column::create(columns[0]->size(), 0);
+        return Int64Column::create(columns[0]->size(), 0);
     }
 
     const auto& array_column = extract_array_column(columns[0].get());
